kõik korras

+++ Model/Ticket.cs
@@ -5,8 +5,7 @@
         public int Id { get; set; }
         public int SessionId { get; set; }
         public string SeatNo { get; set; }
-        public int Price { get; set; }
-
+        public float Price { get; set; }
 		public Session? Session { get; set; }
 	}
 }
